Clarify NoteRequest return when sending 'cmd'
Clarify in function comments that a `true` return value from
NoteRequest when sending a `cmd` indicates that the `cmd`
was sent without error, but does not guarantee that the Notecard
received and processed the `cmd`.
The application should make other provisions to determine that
the `cmd` was processed if this is critical to system behavior.