Add unit tests to cover n_ua.c.
Also, fix a couple problems in n_ua.c:

- The logic for putting the Atollic GCC compiler in the user agent object was
written such that it would never trigger.
- We allow the user to set n_cpu_name via NoteSetUserAgentCPU, but then in
NoteUserAgent we were overriding whatever they set it to, based on preprocessor
logic. I moved that preprocessor logic out so that it controls the default
value of n_cpu_name, but if the user overrides it, that's what we put in the
user agent object.