Don't bother weakening mocked symbols for unit tests.
Instead, force note-c to always be built as a shared library. Then, when the
test binaries are built, the mocks for any note-c functions supersede any of
their "real" counterparts in the shared library. This means that the mock will
always be called and not the "real" function. See:
https://stackoverflow.com/a/4434602/3646689

The reason we were weakening the mocked symbols in the first place is that you
have to if you build note-c as a static lib. If you don't, the mocks will cause
multiple definition errors. If you weaken the function being mocked, though, the
mock is able to override the "real" function. Forcing note-c to be a shared lib
is an easier solution.