Fix tunnelmole step in HIL tests.
Prior to this commit, tunnelmole was launched in the background and the GitHub
job would proceed once the tmole.log file existed. The problem is that the next
step depends on the environment variable MD5SRV_URL being set, and tmole.log
gets created before that gets set. The result is that the next step may run
prematurely, before MD5SRV_URL is set, causing that step to fail. The solution
is to use a "ready" file. The tunnelmole step doesn't proceed until this file
is created. run_tunnelmole.sh only creates this file once it has done
everything it needs to, which includes setting MD5SRV_URL.