Add a CI build that checks for unpermitted libc functions.
After some discussion with Ray, we settled on certain libc functions that we're
ok using in note-c. Everything else from libc should be excluded. This commit
adds a CMake option to build the note-c library without libc. It also tells the
linker to generate errors for undefined references. The result is that we can
run this build to see what libc functions we're using in note-c. A new Bash
script, check_libc_dependencies.sh, processes the output of this build. If any
non-permitted functions are found in the undefined reference errors, the script
fails. This protects us against the introduction of non-permitted libc
functions. A new job in ci.yml runs this script on every PR and push to master.