Fix NOTE_LOWMEM.
Prior to this commit, NOTE_LOWMEM was fundamentally broken because it would
cause our core transaction logic to never return a response from the Notecard.
This commit fixes that and the relevant unit tests. There were some other unit
tests that were broken for different reasons. For example, the CRC tests should
never run with NOTE_LOWMEM defined because all the CRC logic is cut out.

Additionally, the n_ua.c is empty if NOTE_LOWMEM is defined. This implies that
NOTE_DISABLE_USER_AGENT should be defined if NOTE_LOWMEM is defined, so I added
that logic to n_lib.h.

This commit also adds a CI build that runs the unit tests with NOTE_LOWMEM
defined.