Improve cJSON number handling. (#132)
- Get rid of CJSON_NO_CLIB and its associated code paths, per Ray's request.
- Change the type of the J object's valueint member. It should be int32_t if
NOTE_LOWMEM and int64_t otherwise. Capture this in a new #define for the type,
JINTEGER.
- When parsing a JSON number from a JSON string, if the number is within the
bounds of the JINTEGER type, convert it to an integer using JAtoI rather than
simply casting valuenumber to JINTEGER, which can lose precision (e.g. for a
Unix timestamp). If the number is outside the bounds of a JINTEGER, saturate,
which is the same behavior as before this commit.
- When printing a JSON number, print it as an integer if possible. Otherwise,
print it as floating point.
- Fix an issue where JItoA would fail to convert the minimum long int value
(i.e. LONG_MIN) to a string.
- Add a slew of new unit tests to make sure number handling behaves as expected.