Add documentation for a few additional cJSON functions.
`JPrintUnformatted` is particularly important because it's what allows the user
to convert a `J *` to a string. We already document `JParse`, which does the
reverse, so we should document `JPrintUnformatted`, too.

Because the function dynamically allocates memory, we need to tell the user how
to free it. So, I also added documentation for `JFree`. And while I was at it,
I added documentation for `JMalloc`.