Make more JSON number handling improvements.
- Incorporate Ray's improvements to n_ftoa.c and n_atof.c that allow us to
support larger exponents.
- Add JAtoN and JNtoA unit tests to test these improvements.
- If NOTE_C_LOW_MEM is defined, define JNUMBER to float. On truly low memory
platforms, this likely does nothing, because those platforms often map double
and float to the same underlying type (which is single-precision). However, when
running the unit tests on a CPU with double-precision floating point support,
this allows us to deliberately run the unit tests with JNUMBER set to
single-precision. Doing this broke a couple test cases in
JSON_number_handling_test.cpp, which I've fixed here.