Refactor card.binary HIL GitHub workflow.
- Remove the unused localtunnel code. If we ever need to switch to localtunne
from tunnelmole, we can pull it out of the git history.
- Remove usage of third party action pyTooling/Actions/with-post-step. Instead,
to run cleanup steps, use the `if: always()` construct.
- Remove usage of third party action JarvusInnovations/background-action@v1.
The scripts we need to run in the background can simply be launched in the
background using & syntax.
- Refactor a bunch of conditional logic that was hard to read. Instead of using
if statements, we were chaining together a bunch of different Bash statements
using logical operators. Using if statements is simpler to read and understand
what exactly is being checked.
- Refactor run_tunnelmole.sh.