fix: two fixes to improve performance of notepod
First, the method NotecardRequestResponseJSON had a very bad behavior when there was an I/O error, returning NULL instead of a JSON object that would indicate the source of the error.

second, for both standard NotecardRequestResponse as well as NotecardRequestResponseJSON, if there is an error the 'error object' should include the "id" field of the requestor, for tracking of requests/responses.