NF-613: Fix card.binary HIL tests.
- Hardcode Notestation to los-angeles-notestation-1. This one's known to be set
up for running these tests. I have work in progress to make it so this doesn't
need to be hardcoded (i.e. the Notestation client can discover which stations
can run these tests), but it's not finished, yet.
- Make sure to kill the Notestation client in the GitHub Actions workflow when
done with the Notestation.
- Move the `file .pio/build/debug/firmware.elf` gdb command earlier in the
custom upload command for PlatformIO. This just gets rid of some warnings I was
seeing in the log. No functional difference.
- Change test_port to /tmp/ns_mcu_usb from /tmp/mcu_usb. The former is the new
default path for the MCU USB device with the latest Notestation client.