fix: do not wipe response structure on bad bin error AND lock I/O during bin transfer (#181)
* fix: do not wipe response structure on bad bin error

When there is a bad bin error in card.binary, the response looks something like this:
{"req":"card.binary","crc":"000A:21904900"}
{"err":"binary receive prematurely terminated (15062 more bytes expected) {bad-bin}{io}","connected":true,"max":130554,"crc":"000A:521E8DA7"}

In n_helpers.c NoteBinaryStoreTransmit(), this cound in fact be the case in the card.binary call made at L533.

However, as you can see, by following the flow, the c_badbinerr is ignored (as it should be) on L545, and we then extract the "max" and "length" fields from the response on L556.  You would think that everything would proceed just fine.

HOWEVER, because of a bug in n_request.c, this is not the case.  The way that it was coded, if there was a c_badbinerr, n_request.c would WIPE OUT the contents of the rsp and replace it with an errDoc(), thus losing the "max" and "length" and "connected" fields altogether.

This PR fixes that problem.

* lock i2c around chunk I/O during binary transfer
* add card.binary reset for robustness
* allow for concurrency when swapping hooks
* feat: NoteSetActiveInterface
* chore: Create test for new locking behavior
* chore: Cover n_hooks.c

---------

Co-authored-by: Zachary J. Fields <zachary_fields@yahoo.com>